Goodbye alerts; hello text boxes
Original solution was 100% through prompts, alerts, and confirmations AKA annoying pop-ups.  With this commit those are all eliminated, and I/O is handled via text and button elements on the page itself.  The button should only be active when the text boxes contain valid values, and WebStorage auto-fills the textboxes if content has been saved.  The submit button also automatically saves the content.